+++ flink-runtime/src/main/java/org/apache/flink/runtime/executiongraph/restart/RestartStrategyFactory.java
@@ -82,38 +82,41 @@ public abstract class RestartStrategyFactory implements Serializable {
 	 * @throws Exception which indicates that the RestartStrategy could not be instantiated.
 	 */
 	public static RestartStrategyFactory createRestartStrategyFactory(Configuration configuration) throws Exception {
-		String restartStrategyName = configuration.getString(ConfigConstants.RESTART_STRATEGY, "none");
+		String restartStrategyName = configuration.getString(ConfigConstants.RESTART_STRATEGY, null);
+
+		if (restartStrategyName == null) {
+			// support deprecated ConfigConstants values
+			final int numberExecutionRetries = configuration.getInteger(ConfigConstants.RESTART_STRATEGY_FIXED_DELAY_ATTEMPTS,
+				ConfigConstants.DEFAULT_EXECUTION_RETRIES);
+			String pauseString = configuration.getString(AkkaOptions.WATCH_HEARTBEAT_PAUSE);
+			String delayString = configuration.getString(ConfigConstants.RESTART_STRATEGY_FIXED_DELAY_DELAY,
+				pauseString);
+
+			long delay;
+
+			try {
+				delay = Duration.apply(delayString).toMillis();
+			} catch (NumberFormatException nfe) {
+				if (delayString.equals(pauseString)) {
+					throw new Exception("Invalid config value for " +
+						AkkaOptions.WATCH_HEARTBEAT_PAUSE.key() + ": " + pauseString +
+						". Value must be a valid duration (such as '10 s' or '1 min')");
+				} else {
+					throw new Exception("Invalid config value for " +
+						ConfigConstants.RESTART_STRATEGY_FIXED_DELAY_DELAY + ": " + delayString +
+						". Value must be a valid duration (such as '100 milli' or '10 s')");
+				}
+			}
+
+			if (numberExecutionRetries > 0 && delay >= 0) {
+				return new FixedDelayRestartStrategy.FixedDelayRestartStrategyFactory(numberExecutionRetries, delay);
+			} else {
+				return new NoOrFixedIfCheckpointingEnabledRestartStrategyFactory();
+			}
+		}
 
 		switch (restartStrategyName.toLowerCase()) {
 			case "none":
-				// support deprecated ConfigConstants values
-				final int numberExecutionRetries = configuration.getInteger(ConfigConstants.RESTART_STRATEGY_FIXED_DELAY_ATTEMPTS,
-					ConfigConstants.DEFAULT_EXECUTION_RETRIES);
-				String pauseString = configuration.getString(AkkaOptions.WATCH_HEARTBEAT_PAUSE);
-				String delayString = configuration.getString(ConfigConstants.RESTART_STRATEGY_FIXED_DELAY_DELAY,
-					pauseString);
-
-				long delay;
-
-				try {
-					delay = Duration.apply(delayString).toMillis();
-				} catch (NumberFormatException nfe) {
-					if (delayString.equals(pauseString)) {
-						throw new Exception("Invalid config value for " +
-							AkkaOptions.WATCH_HEARTBEAT_PAUSE.key() + ": " + pauseString +
-							". Value must be a valid duration (such as '10 s' or '1 min')");
-					} else {
-						throw new Exception("Invalid config value for " +
-							ConfigConstants.RESTART_STRATEGY_FIXED_DELAY_DELAY + ": " + delayString +
-							". Value must be a valid duration (such as '100 milli' or '10 s')");
-					}
-				}
-
-				if (numberExecutionRetries > 0 && delay >= 0) {
-					return new FixedDelayRestartStrategy.FixedDelayRestartStrategyFactory(numberExecutionRetries, delay);
-				} else {
-					return NoRestartStrategy.createFactory(configuration);
-				}
 			case "off":
 			case "disable":
 				return NoRestartStrategy.createFactory(configuration);
@@ -149,7 +152,7 @@ public abstract class RestartStrategyFactory implements Serializable {
 				}
 
 				// fallback in case of an error
-				return NoOrFixedIfCheckpointingEnabledRestartStrategy.createFactory(configuration);
+				return new NoOrFixedIfCheckpointingEnabledRestartStrategyFactory();
 		}
 	}
 }
